Hide the flow bridge implementations
