+++ libcaf_net/caf/detail/ws_flow_bridge.hpp
@@ -4,68 +4,24 @@
 
 #pragma once
 
-#include "caf/net/web_socket/frame.hpp"
-#include "caf/net/web_socket/lower_layer.hpp"
+#include "caf/net/fwd.hpp"
 #include "caf/net/web_socket/upper_layer.hpp"
 
-#include "caf/async/consumer_adapter.hpp"
-#include "caf/async/producer_adapter.hpp"
-#include "caf/async/spsc_buffer.hpp"
-#include "caf/detail/flow_bridge_base.hpp"
+#include "caf/async/fwd.hpp"
+#include "caf/detail/ws_conn_acceptor.hpp"
 #include "caf/fwd.hpp"
-#include "caf/sec.hpp"
 
-#include <utility>
+#include <memory>
 
 namespace caf::detail {
 
-/// Convenience alias for referring to the base type of @ref flow_bridge.
-template <class Base>
-using ws_flow_bridge_base_t
-  = detail::flow_bridge_base<Base, net::web_socket::lower_layer,
-                             net::web_socket::frame>;
+CAF_NET_EXPORT
+std::unique_ptr<net::web_socket::upper_layer>
+make_ws_flow_bridge(async::consumer_resource<net::web_socket::frame> pull,
+                    async::producer_resource<net::web_socket::frame> push);
 
-/// Translates between a message-oriented transport and data flows.
-template <class Base>
-class ws_flow_bridge : public ws_flow_bridge_base_t<Base> {
-public:
-  using super = ws_flow_bridge_base_t<Base>;
-
-  using super::super;
-
-  bool write(const net::web_socket::frame& item) override {
-    if (item.is_binary()) {
-      super::down_->begin_binary_message();
-      auto& bytes = super::down_->binary_message_buffer();
-      auto src = item.as_binary();
-      bytes.insert(bytes.end(), src.begin(), src.end());
-      return super::down_->end_binary_message();
-    } else {
-      super::down_->begin_text_message();
-      auto& text = super::down_->text_message_buffer();
-      auto src = item.as_text();
-      text.insert(text.end(), src.begin(), src.end());
-      return super::down_->end_text_message();
-    }
-  }
-
-  // -- implementation of web_socket::lower_layer ------------------------------
-
-  ptrdiff_t consume_binary(byte_span buf) override {
-    if (!super::out_)
-      return -1;
-    if (super::out_.push(net::web_socket::frame{buf}) == 0)
-      super::down_->suspend_reading();
-    return static_cast<ptrdiff_t>(buf.size());
-  }
-
-  ptrdiff_t consume_text(std::string_view buf) override {
-    if (!super::out_)
-      return -1;
-    if (super::out_.push(net::web_socket::frame{buf}) == 0)
-      super::down_->suspend_reading();
-    return static_cast<ptrdiff_t>(buf.size());
-  }
-};
+CAF_NET_EXPORT
+std::unique_ptr<net::web_socket::upper_layer::server>
+make_ws_flow_bridge(detail::ws_conn_acceptor_ptr wca);
 
 } // namespace caf::detail
